requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
Werkzeug==3.0.3
